+++ gvm/protocols/gmpv208/__init__.py
@@ -28,7 +28,7 @@ Module for communication with gvmd in
 from typing import Any, Callable, Optional
 
 from gvm.protocols.gmpv208.gmpv208 import GmpV208Mixin
-from gvm.protocols.entities.tasks import TaskMixin
+from gvm.protocols.gmpv208.entities.tasks import TaskMixin
 from gvm.connections import GvmConnection
